Extract helpers from calculate treatments spec
