+++ shared/tests/utils/catering.spec.ts
@@ -1,0 +1,54 @@
+import produce from 'immer';
+import type { ExerciseState } from '../../src/state';
+import { getElement } from '../../src/store/action-reducers/utils';
+import type { UUID } from '../../src/utils';
+
+export interface Catering {
+    /**
+     * The id of the material or personnel catering
+     */
+    catererId: UUID;
+    catererType: 'material' | 'personnel';
+    /**
+     * All patients treated by {@link catererId}
+     */
+    patientIds: UUID[];
+}
+
+/**
+ * Asserts that adding the specified {@link caterings} to the {@link beforeState} results in the {@link newState}.
+ * If the {@link beforeState} has already caterings in it, these will not be removed.
+ */
+export function assertCatering(
+    beforeState: ExerciseState,
+    newState: ExerciseState,
+    caterings: Catering[]
+) {
+    // Add all caterings to the before state and look whether the result is the newState
+    const shouldState = produce(beforeState, (draftState) => {
+        for (const catering of caterings) {
+            // Update all the patients
+            const patients = catering.patientIds.map((patientId) =>
+                getElement(draftState, 'patient', patientId)
+            );
+            for (const patient of patients) {
+                patient[
+                    catering.catererType === 'material'
+                        ? 'assignedMaterialIds'
+                        : 'assignedPersonnelIds'
+                ][catering.catererId] = true;
+            }
+            // Update the catering element
+            const cateringElement = getElement(
+                draftState,
+                catering.catererType,
+                catering.catererId
+            );
+            for (const patientId of catering.patientIds) {
+                cateringElement.assignedPatientIds[patientId] = true;
+            }
+        }
+        return draftState;
+    });
+    expect(newState).toStrictEqual(shouldState);
+}
