+++ shared/tsconfig.build.json
@@ -1,5 +1,8 @@
 {
     "extends": "./tsconfig.json",
+    "compilerOptions": {
+        "rootDir": "./src"
+    },
     "include": ["src/**/*"],
     "exclude": ["node_modules", "src/**/*.spec.ts"]
 }
